The whole Python backend has been reworked and so there were fixed loads of errors, like the crash on syncing if using the newest python version. Stability has improved too and some memory leaks were fixed

+++ hnpython/src/pyargs/pyargs.cpp
@@ -3,7 +3,7 @@
 
 #include "log.h"
 
-PyArgs::PyArgs(size_t argc){
+PyArgs::PyArgs(int argc){
     FUN();
     this->_argc = argc;
     this->_argv = (void*)PyTuple_New(this->_argc);
